Move some utility test code around, clean up
Cleaned up some tests to use the wrapper macros (more ergonomic)
and deleted some deprecated code.
