+++ base_layer/core/src/test_utils/primitives.rs
@@ -1,4 +1,4 @@
-// Copyright 2019 The Tari Project
+// Copyright 2019. The Tari Project
 //
 // Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
 // following conditions are met:
@@ -19,5 +19,39 @@
 // SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
 
-pub mod simple_block_chain;
+use crate::{
+    tari_amount::MicroTari,
+    transaction_protocol::{build_challenge, TransactionMetadata},
+    types::{PrivateKey, PublicKey, Signature},
+};
+use tari_crypto::keys::{PublicKey as PK, SecretKey as SK};
+
+/// A convenience struct for a set of public-private keys and a public-private nonce
+pub struct TestKeySet {
+    pub k: PrivateKey,
+    pub pk: PublicKey,
+    pub r: PrivateKey,
+    pub pr: PublicKey,
+}
+
+/// Generate a new random key set. The key set includes
+/// * a public-private keypair (k, pk)
+/// * a public-private nonce keypair (r, pr)
+pub fn generate_keys() -> TestKeySet {
+    let mut rng = rand::OsRng::new().unwrap();
+    let (k, pk) = PublicKey::random_keypair(&mut rng);
+    let (r, pr) = PublicKey::random_keypair(&mut rng);
+    TestKeySet { k, pk, r, pr }
+}
+
+/// Generate a random transaction signature, returning the public key (excess) and the signature.
+pub fn create_random_signature(fee: MicroTari, lock_height: u64) -> (PublicKey, Signature) {
+    let mut rng = rand::OsRng::new().unwrap();
+    let r = PrivateKey::random(&mut rng);
+    let (k, p) = PublicKey::random_keypair(&mut rng);
+    let tx_meta = TransactionMetadata { fee, lock_height };
+    let e = build_challenge(&PublicKey::from_secret_key(&r), &tx_meta);
+    (p, Signature::sign(k, r, &e).unwrap())
+}
